fix deleting intervals
