+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon ./src/main.js",
-    "start": "node ./src/main.js"
+    "dev": "nodemon main.js",
+    "start": "node main.js"
   },
   "keywords": [],
   "author": "",
