+++ src/withIntervalSMS.js
@@ -1,5 +1,5 @@
 import config from "config";
-import {promptForGptAsk} from "./main.js";
+import {promptForGptAsk} from "../main.js";
 
 const firstPart = config.get('GPT_TEMPLATE_1');
 const secondPart = config.get('GPT_TEMPLATE_2');
@@ -8,7 +8,7 @@ export function withIntervalSMS(ctx, message,  intervalMinutes) {
     ${ctx.from.first_name} ${ctx.from.last_name}`
     return setInterval(async () => {
         await promptForGptAsk(ctx,firstPart + message + secondPart + lastPart);
-    }, intervalMinutes * 60000);
+    }, intervalMinutes * 10000);
 }
